# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/gift": {
            "post": {
                "description": "Initialize a new gift.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GiftDTO"
                ],
                "summary": "Initialize gift",
                "parameters": [
                    {
                        "description": "Gift init request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gift.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gift.DTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/gift/use/{giftCode}": {
            "post": {
                "description": "Use a gift code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GiftDTO"
                ],
                "summary": "Use gift",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gift code",
                        "name": "giftCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gift.DTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/gift/{giftCode}": {
            "get": {
                "description": "Get a gift by code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GiftDTO"
                ],
                "summary": "Get gift",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gift code",
                        "name": "giftCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gift.DTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Health check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "gift.CreateRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "codePrefix": {
                    "type": "string"
                },
                "expirationDate": {
                    "type": "string"
                },
                "giftAmount": {
                    "type": "integer"
                },
                "startDateTime": {
                    "type": "string"
                },
                "usageLimit": {
                    "type": "integer"
                }
            }
        },
        "gift.DTO": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "expirationDate": {
                    "type": "string"
                },
                "giftAmount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "startDateTime": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "usageLimit": {
                    "type": "integer"
                },
                "usedCount": {
                    "type": "integer"
                }
            }
        },
        "handler.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/serr.ErrorCode"
                },
                "message": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
        "serr.ErrorCode": {
            "type": "string",
            "enum": [
                "INTERNAL",
                "INVALID_GIFT_ID",
                "INVALID_GIFT_CODE",
                "INVALID_DISCOUNT_ID",
                "INVALID_DISCOUNT_CODE",
                "PERMISSION",
                "GIFT_USAGE_LIMIT_REACHED"
            ],
            "x-enum-varnames": [
                "ErrInternal",
                "ErrInvalidGiftID",
                "ErrInvalidGiftCode",
                "ErrInvalidDiscountID",
                "ErrInvalidDiscountCode",
                "ErrPermission",
                "ErrGiftUsageLimitReached"
            ]
        }
    }
}